refactor line edit
