--- tabulous/_qt/_table/_base/_header_line_edit.py
@@ -1,122 +1,0 @@
-from __future__ import annotations
-from typing import TYPE_CHECKING
-from qtpy import QtCore
-from qtpy.QtCore import Qt
-
-import pandas as pd
-from ._line_edit import QTableLineEdit
-
-from ....types import HeaderInfo
-
-if TYPE_CHECKING:
-    from qtpy.QtCore import pyqtBoundSignal
-    from ._header_view import QDataFrameHeaderView
-    from ._table_base import QMutableTable
-
-
-class _QHeaderLineEdit(QTableLineEdit):
-    """Line edit used for editing header text."""
-
-    ALIGNMENT: Qt.AlignmentFlag
-
-    def _get_index(self) -> int:
-        raise NotImplementedError()
-
-    def _get_rect(self, index: int) -> QtCore.QRect:
-        raise NotImplementedError()
-
-    def _get_pandas_axis(self) -> pd.Index:
-        raise NotImplementedError()
-
-    def _get_signal(self) -> pyqtBoundSignal:
-        raise NotImplementedError()
-
-    def __init__(
-        self,
-        parent: QDataFrameHeaderView,
-        table: QMutableTable | None = None,
-        pos: tuple[int, int] = (0, 0),
-    ):
-        super().__init__(parent, table, pos)
-
-        index = self._get_index()
-        geometry = self._get_rect(index)
-        geometry.adjust(2, 1, -2, -1)
-        self.setGeometry(geometry)
-        self.setAlignment(self.ALIGNMENT)
-        self.setHidden(False)
-
-        df_axis = self._get_pandas_axis()
-        if index < df_axis.size:
-            old_value = df_axis[index]
-            text = str(old_value)
-        else:
-            # spreadsheet
-            old_value = None
-            text = ""
-
-        @self.editingFinished.connect
-        def _set_header_data():
-            self.editingFinished.disconnect()
-            value = self.text()
-            if not value == text:
-                self._get_signal().emit(HeaderInfo(index, value, text))
-            table = self._table._qtable_view
-            table.setFocus()
-            self.deleteLater()
-            return None
-
-        self.selectAll()
-        self.setFocus()
-
-    def isTextValid(self, r: int, c: int, text: str) -> bool:
-        """True if text is valid for this cell."""
-        pd_index = self._get_pandas_axis()
-        idx = self._get_index()
-        not_in = text not in pd_index
-        if idx < pd_index.size:
-            return text == pd_index[idx] or not_in
-        else:
-            return not_in
-
-
-class QVerticalHeaderLineEdit(_QHeaderLineEdit):
-    ALIGNMENT = Qt.AlignmentFlag.AlignLeft
-
-    def _get_index(self) -> int:
-        return self._pos[0]
-
-    def _get_rect(self, index: int) -> QtCore.QRect:
-        header = self._table._qtable_view.verticalHeader()
-        width = header.width()
-        height = header.sectionSize(index)
-        left = header.rect().left()
-        top = header.sectionViewportPosition(index)
-        return QtCore.QRect(left, top, width, height)
-
-    def _get_pandas_axis(self) -> pd.Index:
-        return self._table.model().df.index
-
-    def _get_signal(self):
-        return self._table.rowChangedSignal
-
-
-class QHorizontalHeaderLineEdit(_QHeaderLineEdit):
-    ALIGNMENT = Qt.AlignmentFlag.AlignCenter
-
-    def _get_index(self) -> int:
-        return self._pos[1]
-
-    def _get_rect(self, index: int) -> QtCore.QRect:
-        header = self._table._qtable_view.horizontalHeader()
-        width = header.sectionSize(index)
-        height = header.height()
-        left = header.sectionViewportPosition(index)
-        top = header.rect().top()
-        return QtCore.QRect(left, top, width, height)
-
-    def _get_pandas_axis(self) -> pd.Index:
-        return self._table.model().df.columns
-
-    def _get_signal(self):
-        return self._table.columnChangedSignal
